Migrate `cypress` to v10
