+++ .gitignore
@@ -42,7 +42,6 @@ cypress/screenshots
 cypress/videos
 cypress/plugins/index.js
 cypress/plugins/cy-ts-preprocessor.js
-cypress/tsconfig.json
 
 # GO
 /vendor
