--- cypress.json
@@ -1,15 +1,0 @@
-{
-  "baseUrl": "http://localhost:8000",
-  "supportFile": "./cypress/support/index.ts",
-  "chromeWebSecurity": false,
-  "screenshotOnRunFailure": true,
-  "videoCompression": false,
-  "pageLoadTimeout": 60000,
-  "defaultCommandTimeout": 60000,
-  "responseTimeout": 60000,
-  "requestTimeout": 60000,
-  "viewportHeight": 1080,
-  "viewportWidth": 1920,
-  "projectId": "mdsuba",
-  "video": true
-}
