+++ cypress/e2e/providers/azure.spec.ts
