+++ cypress/e2e/providers/digitalocean.spec.ts
