+++ cypress/e2e/providers/equinix.spec.ts
