+++ cypress/e2e/providers/kubeadm.spec.ts
