+++ cypress/e2e/providers/nutanix.spec.ts
