+++ cypress/e2e/providers/vsphere.spec.ts
