+++ cypress/e2e/stories/admin-settings/cluster-settings.spec.ts
