+++ cypress/e2e/stories/admin-settings/machine-deployment-replicas.spec.ts
