+++ cypress/e2e/stories/admin-settings/presets.spec.ts
