+++ cypress/e2e/stories/admin-settings/resource-quota.spec.ts
