+++ cypress/e2e/stories/cluster-templates.spec.ts
