+++ cypress/e2e/stories/external-cluster.spec.ts
