+++ cypress/e2e/v2/stories/admin-settings/defaults.spec.ts
