+++ cypress/e2e/v2/stories/edition.spec.ts
