+++ cypress/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "target": "ES2019",
     "strict": true,
     "baseUrl": "./",
     "paths": {
@@ -15,5 +16,6 @@
   },
   "include": [
     "**/*.ts",
+    "../cypress.config.ts"
   ]
 }
