+++ package.json
@@ -39,7 +39,7 @@
     "fix:license": "license-check-and-add add -r 2022",
     "clean": "gts clean",
     "vi": "node version.js",
-    "postinstall": "ngcc && rm -rf cypress/**/*.js && rm -f cypress/tsconfig.json",
+    "postinstall": "ngcc",
     "prepare": "husky install"
   },
   "lint-staged": {
@@ -98,8 +98,6 @@
     "@angular-devkit/build-angular": "14.2.2",
     "@angular/cli": "14.2.2",
     "@angular/compiler-cli": "14.2.1",
-    "@bahmutov/add-typescript-to-cypress": "2.1.2",
-    "@cypress/webpack-preprocessor": "5.11.1",
     "@jest/globals": "29.0.3",
     "@types/jest": "29.0.2",
     "@types/js-yaml": "4.0.5",
@@ -110,9 +108,9 @@
     "browserlist": "1.0.1",
     "btoa": "1.2.1",
     "concurrently": "7.4.0",
-    "cypress": "9.7.0",
-    "cypress-fail-fast": "3.4.1",
-    "del": "6.1.1",
+    "cypress": "10.8.0",
+    "cypress-fail-fast": "5.0.1",
+    "del": "7.0.0",
     "eslint": "8.23.1",
     "eslint-plugin-compat": "4.0.2",
     "eslint-plugin-prettier": "4.2.1",
