+++ src/app/shared/components/add-project-dialog/template.html
@@ -39,7 +39,7 @@ limitations under the License.
     </form>
   </mat-dialog-content>
   <mat-dialog-actions>
-    <km-button id="km-add-project-dialog-save"
+    <km-button buttonId="km-add-project-dialog-save"
                icon="km-icon-save"
                label="Save Project"
                [disabled]="!form.valid"
